engine: extract esm/place.js
